separate addon markers from runtime helpers
