+++ addons/api/api_test.go
@@ -1,8 +1,6 @@
 package api
 
 import (
-	"net/http"
-	"net/http/httptest"
 	"testing"
 
 	"github.com/cssbruno/gowdk"
@@ -15,20 +13,5 @@ func TestAddonRegistersAPIFeature(t *testing.T) {
 	}
 	if !(gowdk.Config{Addons: []gowdk.Addon{addon}}).HasFeature(gowdk.FeatureAPI) {
 		t.Fatal("expected api feature")
-	}
-}
-
-func TestRuntimeReExports(t *testing.T) {
-	registry := Registry{
-		"health": func(writer http.ResponseWriter, request *http.Request) {
-			writer.WriteHeader(http.StatusNoContent)
-		},
-	}
-
-	recorder := httptest.NewRecorder()
-	registry["health"](recorder, httptest.NewRequest(http.MethodGet, "/api/health", nil))
-
-	if recorder.Code != http.StatusNoContent {
-		t.Fatalf("unexpected status: %d", recorder.Code)
 	}
 }
