+++ addons/partial/partial_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/cssbruno/gowdk"
-	"github.com/cssbruno/gowdk/runtime/response"
 )
 
 func TestAddonRegistersPartialFeature(t *testing.T) {
@@ -14,12 +13,5 @@ func TestAddonRegistersPartialFeature(t *testing.T) {
 	}
 	if !(gowdk.Config{Addons: []gowdk.Addon{addon}}).HasFeature(gowdk.FeaturePartial) {
 		t.Fatal("expected partial feature")
-	}
-}
-
-func TestRuntimeReExports(t *testing.T) {
-	result := Fragment("#messages", "<li>Hello</li>")
-	if result.Kind != response.Fragment || result.Target != "#messages" || result.Body != "<li>Hello</li>" {
-		t.Fatalf("unexpected fragment response: %#v", result)
 	}
 }
