+++ addons/ratelimit/ratelimit_test.go
@@ -2,7 +2,6 @@ package ratelimit
 
 import (
 	"testing"
-	"time"
 
 	"github.com/cssbruno/gowdk"
 )
@@ -14,19 +13,5 @@ func TestAddonRegistersRateLimitFeature(t *testing.T) {
 	}
 	if !(gowdk.Config{Addons: []gowdk.Addon{addon}}).HasFeature(gowdk.FeatureRateLimit) {
 		t.Fatal("expected ratelimit feature")
-	}
-}
-
-func TestRuntimeReExports(t *testing.T) {
-	limiter, err := New(Options{
-		Limit:  1,
-		Window: time.Minute,
-		Store:  NewInMemoryStore(InMemoryOptions{}),
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if limiter == nil {
-		t.Fatal("expected limiter")
 	}
 }
