+++ internal/compiler/backend_typed_signatures.go
@@ -211,7 +211,6 @@ func isTypedPointerToNamed(typ types.Type, importPath, name string) bool {
 
 func isTypedLoadContext(typ types.Type) bool {
 	return isTypedNamed(typ, ssrImportPath, "LoadContext") ||
-		isTypedNamed(typ, runtimeSSRImportPath, "LoadContext") ||
 		isTypedNamed(typ, guardImportPath, "Context")
 }
